+++ include/agb_flash.h
@@ -7,8 +7,8 @@
 
 // Exported ROM declarations
 
-u16 SetFlashTimerIntr(u8 timerNum, void (**intrFunc)(void));
-u16 IdentifyFlash(void);
-u32 ProgramFlashSectorAndVerify(u16 sectorNum, u8 *src);
+u32 SetFlashTimerIntr(u8 timerNum, void (**intrFunc)(void));
+u32 IdentifyFlash(void);
+u32 ProgramFlashSectorAndVerify(u32 sectorNum, u8 *src);
 
 #endif //GUARD_AGB_FLASH_H
