+++ include/battle_dome.h
@@ -1,7 +1,7 @@
 #ifndef GUARD_BATTLE_DOME_H
 #define GUARD_BATTLE_DOME_H
 
-int GetDomeTrainerSelectedMons(u16 tournamentTrainerId);
-int TrainerIdToDomeTournamentId(u16 trainerId);
+int GetDomeTrainerSelectedMons(u32 tournamentTrainerId);
+int TrainerIdToDomeTournamentId(u32 trainerId);
 
 #endif // GUARD_BATTLE_DOME_H
